+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Added
 
 - Fields of a `def` defined in terms of each other, e.g. `int g = f; let f = g;`, are reported as an error.
+- A class statement following a definition of the same class, be it in the same file or a file included before, is
+  reported as an error. A quick fix navigates to the closest such definition.
 
 ### Changed